Turn react bootloader into typescript, and refactor.
